Addition of Pictures
Finding it difficult becoming accustomed to Git and GitHub. Spent too long thinking there was a complicated way to insert the photos and learned there was an upload spot. I have no idea how to size them. On the screen shot it says "width" but not in the downloaded image of Groot.